extract StoreSettingsViewModel
