+++ WooCommerce/Classes/POS/Presentation/Settings/PointOfSaleSettingsStoreDetailView.swift
@@ -1,84 +1,14 @@
 import SwiftUI
-import Yosemite
-
-final class StoreSettingsViewModel: ObservableObject {
-    @Published var receiptInformation = POSReceiptInformation.empty
-    @Published var shouldShowReceiptInformation: Bool = false
-
-    private let siteID: Int64
-    private let settingsService: PointOfSaleSettingsServiceProtocol
-    private let pluginsService: PluginsServiceProtocol
-
-    init(siteID: Int64,
-         settingsService: PointOfSaleSettingsServiceProtocol,
-         pluginsService: PluginsServiceProtocol) {
-        self.siteID = siteID
-        self.settingsService = settingsService
-        self.pluginsService = pluginsService
-    }
-
-    @MainActor
-    func retrievePOSReceiptSettings() async {
-        shouldShowReceiptInformation = await isPluginSupported(.wooCommerce, minimumVersion: Constants.minimumWooCommerceVersion)
-
-        guard shouldShowReceiptInformation else {
-            return
-        }
-        do {
-            let siteSettings = try await settingsService.retrievePointOfSaleSettings()
-            updateReceiptSettings(from: siteSettings)
-        } catch {
-            DDLogError("Failed to load POS settings: \(error)")
-        }
-    }
-
-    @MainActor
-    private func isPluginSupported(_ plugin: Plugin,
-                                   minimumVersion: String) async -> Bool {
-        guard let systemPlugin = pluginsService.loadPluginInStorage(siteID: siteID, plugin: plugin, isActive: true), systemPlugin.active else {
-            return false
-        }
-
-        let isSupported = VersionHelpers.isVersionSupported(version: systemPlugin.version,
-                                                            minimumRequired: minimumVersion)
-        return isSupported
-    }
-
-    private func updateReceiptSettings(from siteSettings: [SiteSetting]) {
-        receiptInformation = POSReceiptInformation(
-            storeName: settingValue(from: siteSettings, settingID: "woocommerce_pos_store_name"),
-            storeAddress: settingValue(from: siteSettings, settingID: "woocommerce_pos_store_address"),
-            phone: settingValue(from: siteSettings, settingID: "woocommerce_pos_store_phone"),
-            email: settingValue(from: siteSettings, settingID: "woocommerce_pos_store_email"),
-            refundReturnsPolicy: settingValue(from: siteSettings, settingID: "woocommerce_pos_refund_returns_policy")
-        )
-    }
-
-    private func settingValue(from siteSettings: [SiteSetting], settingID: String) -> String? {
-        let value = siteSettings.first { $0.settingID == settingID }?.value
-        return value?.isEmpty == true ? nil : value
-    }
-}
-
-private extension StoreSettingsViewModel {
-    enum Constants {
-        static let minimumWooCommerceVersion: String = "10.0"
-    }
-}
 
 struct PointOfSaleSettingsStoreDetailView: View {
-    @State private var isLoading: Bool = true
-    @StateObject private var storeViewModel: StoreSettingsViewModel
+    @State private var isLoading: Bool = false
 
     let settingsController: PointOfSaleSettingsControllerProtocol
+    let viewModel: StoreSettingsViewModel
 
-    init(settingsController: PointOfSaleSettingsControllerProtocol) {
+    init(settingsController: PointOfSaleSettingsControllerProtocol, viewModel: StoreSettingsViewModel) {
         self.settingsController = settingsController
-        self._storeViewModel = StateObject(wrappedValue: StoreSettingsViewModel(
-            siteID: settingsController.siteID,
-            settingsService: settingsController.settingsService,
-            pluginsService: settingsController.pluginsService
-        ))
+        self.viewModel = viewModel
     }
 
     var body: some View {
@@ -109,41 +39,41 @@ struct PointOfSaleSettingsStoreDetailView: View {
                     VStack(alignment: .leading, spacing: POSPadding.small) {
                         Text(Localization.receiptStoreName)
                             .font(.posBodyMediumRegular())
-                        settingValueView(for: storeViewModel.receiptInformation.storeName)
+                        settingValueView(for: viewModel.receiptInformation.storeName)
                     }
 
                     VStack(alignment: .leading, spacing: POSPadding.small) {
                         Text(Localization.physicalAddress)
                             .font(.posBodyMediumRegular())
-                        settingValueView(for: storeViewModel.receiptInformation.storeAddress)
+                        settingValueView(for: viewModel.receiptInformation.storeAddress)
                     }
 
                     VStack(alignment: .leading, spacing: POSPadding.small) {
                         Text(Localization.phoneNumber)
                             .font(.posBodyMediumRegular())
-                        settingValueView(for: storeViewModel.receiptInformation.phone)
+                        settingValueView(for: viewModel.receiptInformation.phone)
                     }
 
                     VStack(alignment: .leading, spacing: POSPadding.small) {
                         Text(Localization.email)
                             .font(.posBodyMediumRegular())
-                        settingValueView(for: storeViewModel.receiptInformation.email)
+                        settingValueView(for: viewModel.receiptInformation.email)
                     }
 
                     VStack(alignment: .leading, spacing: POSPadding.small) {
                         Text(Localization.refundReturnsPolicy)
                             .font(.posBodyMediumRegular())
-                        settingValueView(for: storeViewModel.receiptInformation.refundReturnsPolicy)
+                        settingValueView(for: viewModel.receiptInformation.refundReturnsPolicy)
                     }
                 } header: {
                     Text(Localization.receiptInformation)
                         .font(.posBodyLargeRegular())
                 }
-                .renderedIf(storeViewModel.shouldShowReceiptInformation)
+                .renderedIf(viewModel.shouldShowReceiptInformation)
             }
             .task {
                 isLoading = true
-                await storeViewModel.retrievePOSReceiptSettings()
+                await viewModel.retrievePOSReceiptSettings()
                 isLoading = false
             }
         }
@@ -266,6 +196,9 @@ private extension PointOfSaleSettingsStoreDetailView {
 
 #if DEBUG
 #Preview {
-    PointOfSaleSettingsStoreDetailView(settingsController: PointOfSaleSettingsPreviewController())
+    PointOfSaleSettingsStoreDetailView(settingsController: PointOfSaleSettingsPreviewController(),
+                                       viewModel: .init(siteID: 123,
+                                                        settingsService: MockPointOfSaleSettingsService(),
+                                                        pluginsService: PluginsServicePreview()))
 }
 #endif
